настроено полное отображение данных постов

+++ iOS_diplom_VK_App/Helpers/VKDateFormater.swift
@@ -13,10 +13,10 @@ final class VKDateFormater: DateFormatter {
     
     let dateFormater = DateFormatter()
     
-    func formateDate(date: Double) -> String {
+    func formateDate(date: Int) -> String {
         dateFormater.locale = Locale(identifier: "ru_RU")
         dateFormater.dateFormat = "d MMM 'в' HH:mm"
-        let date = Date(timeIntervalSince1970: date)
+        let date = Date(timeIntervalSince1970: Double(date))
         let dateTitle = dateFormater.string(from: date)
         return dateTitle
     }
